(pc-23359)[API] refactor: moving delete_suspended_accounts_after_withdrawal_period to pcapi.core.users.commands
