+++ api/tests/core/users/test_api.py
@@ -1673,3 +1673,47 @@ class NotifyUserBeforeDeletionUponSuspensionTest:
         assert len(mails_testing.outbox) == 1
         assert mails_testing.outbox[0].sent_data["params"]["FIRSTNAME"] == user.firstName
         assert mails_testing.outbox[0].sent_data["To"] == user.email
+
+
+@pytest.mark.usefixtures("db_session")
+class GetSuspendedAccountsUponUserRequestSinceTest:
+    def test_get_suspended_upon_user_request_accounts_since(self) -> None:
+        one_week_ago = datetime.datetime.utcnow() - datetime.timedelta(days=7)
+        something = history_factories.SuspendedUserActionHistoryFactory(
+            actionDate=one_week_ago, reason=users_constants.SuspensionReason.UPON_USER_REQUEST
+        )
+
+        # not suspended upon user request: should be ignored
+        history_factories.SuspendedUserActionHistoryFactory(
+            actionDate=one_week_ago, reason=users_constants.SuspensionReason.FRAUD_SUSPICION
+        )
+
+        # suspended less than 5 days ago (see below): should be ignored
+        yesterday = datetime.datetime.utcnow() - datetime.timedelta(days=1)
+        history_factories.SuspendedUserActionHistoryFactory(
+            actionDate=yesterday, reason=users_constants.SuspensionReason.UPON_USER_REQUEST
+        )
+
+        expected_user_ids = {something.userId}
+
+        with assert_num_queries(1):
+            query = users_api.get_suspended_upon_user_request_accounts_since(5)
+            user_ids = {user.id for user in query}
+            assert user_ids == expected_user_ids
+
+    def test_unsuspended_account(self) -> None:
+        """
+        Test that an unsuspended account is ignored, even if the
+        suspension event occurred more than N days ago.
+        """
+        one_week_ago = datetime.datetime.utcnow() - datetime.timedelta(days=7)
+        user = users_factories.UserFactory(isActive=False)
+        history_factories.SuspendedUserActionHistoryFactory(
+            user=user, actionDate=one_week_ago, reason=users_constants.SuspensionReason.FRAUD_SUSPICION
+        )
+
+        yesterday = datetime.datetime.utcnow() - datetime.timedelta(days=1)
+        history_factories.UnsuspendedUserActionHistoryFactory(user=user, actionDate=yesterday)
+
+        with assert_num_queries(1):
+            assert not list(users_api.get_suspended_upon_user_request_accounts_since(5))
